Improved pager class encapsulation and fixed the download form generated by pager form processing.
Also changed the cookies' names to be prepended by the format name instead of some arbitrary template name.  This will allow the saving and retrieval of rows as a cookie when a search is performed.

Files checked in: DataRepo/forms.py DataRepo/templates/DataRepo/search/results/fcirc.html DataRepo/templates/DataRepo/search/results/peakdata.html DataRepo/templates/DataRepo/search/results/peakgroups.html DataRepo/templates/pagination.html DataRepo/views.py DataRepo/pager.py
[skip ci]
